+++ app-modules/item/database/migrations/2025_07_29_000002_create_item_variants_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
             $table->foreignId('item_id')->constrained()->cascadeOnDelete();
             $table->string('name');
             $table->string('sku')->nullable()->unique()->index();
-            $table->decimal('price_adjustment', 10, 2)->default(0); // Can be positive or negative
+            $table->integer('price_adjustment')->default(0); // Can be positive or negative, stored in minor units
             $table->decimal('size_multiplier', 5, 2)->default(1); // For portion sizes
             $table->boolean('is_default')->default(false);
             $table->boolean('is_active')->default(true);
